+++ inc/minirt.h
@@ -83,6 +83,7 @@ void	setup_scene(t_scene *scene);
 void	set_matrix(t_matrix *old_m, t_matrix new_m);
 float	degrees_to_radians(float degrees);
 int		trace_color(t_ray ray, t_scene *scene);
+float	get_closest_hit(t_intersection intersection);
 int		lighting(t_light light, t_ambient ambient, t_intersection intersection,
 		t_ray ray);
